FIX: Use nitransforms for most xfm handling
