+++ nibabies/interfaces/resampling.py
@@ -4,10 +4,8 @@ import asyncio
 import os
 from collections.abc import Callable
 from functools import partial
-from pathlib import Path
 from typing import TypeVar
 
-import h5py
 import nibabel as nb
 import nitransforms as nt
 import numpy as np
@@ -19,11 +17,12 @@ from nipype.interfaces.base import (
     traits,
 )
 from nipype.utils.filemanip import fname_presuffix
-from nitransforms.io.itk import ITKCompositeH5
 from scipy import ndimage as ndi
 from scipy.sparse import hstack as sparse_hstack
 from sdcflows.transform import grid_bspline_weights
 from sdcflows.utils.tools import ensure_positive_cosines
+
+from nibabies.utils.transforms import load_transforms
 
 R = TypeVar('R')
 
@@ -32,95 +31,6 @@ async def worker(job: Callable[[], R], semaphore: asyncio.Semaphore) -> R:
     async with semaphore:
         loop = asyncio.get_running_loop()
         return await loop.run_in_executor(None, job)
-
-
-def load_transforms(xfm_paths: list[Path], inverse: list[bool]) -> nt.base.TransformBase:
-    """Load a series of transforms as a nitransforms TransformChain
-
-    An empty list will return an identity transform
-    """
-    if len(inverse) == 1:
-        inverse *= len(xfm_paths)
-    elif len(inverse) != len(xfm_paths):
-        raise ValueError('Mismatched number of transforms and inverses')
-
-    chain = None
-    for path, inv in zip(xfm_paths[::-1], inverse[::-1], strict=False):
-        path = Path(path)
-        if path.suffix == '.h5':
-            xfm = load_ants_h5(path)
-        else:
-            xfm = nt.linear.load(path)
-        if inv:
-            xfm = ~xfm
-        if chain is None:
-            chain = xfm
-        else:
-            chain += xfm
-    if chain is None:
-        chain = nt.base.TransformBase()
-    return chain
-
-
-FIXED_PARAMS = np.array([
-    193.0, 229.0, 193.0,  # Size
-    96.0, 132.0, -78.0,   # Origin
-    1.0, 1.0, 1.0,        # Spacing
-    -1.0, 0.0, 0.0,       # Directions
-    0.0, -1.0, 0.0,
-    0.0, 0.0, 1.0,
-])  # fmt:skip
-
-
-def load_ants_h5(filename: Path) -> nt.base.TransformBase:
-    """Load ANTs H5 files as a nitransforms TransformChain"""
-    # Borrowed from https://github.com/feilong/process
-    # process.resample.parse_combined_hdf5()
-    #
-    # Changes:
-    #   * Tolerate a missing displacement field
-    #   * Return the original affine without a round-trip
-    #   * Always return a nitransforms TransformChain
-    #
-    # This should be upstreamed into nitransforms
-    h = h5py.File(filename)
-    xform = ITKCompositeH5.from_h5obj(h)
-
-    # nt.Affine
-    transforms = [nt.Affine(xform[0].to_ras())]
-
-    if '2' not in h['TransformGroup']:
-        return transforms[0]
-
-    transform2 = h['TransformGroup']['2']
-
-    # Confirm these transformations are applicable
-    if transform2['TransformType'][:][0] not in (
-        b'DisplacementFieldTransform_float_3_3',
-        b'DisplacementFieldTransform_double_3_3',
-    ):
-        msg = 'Unknown transform type [2]\n'
-        for i in h['TransformGroup'].keys():
-            msg += f'[{i}]: {h["TransformGroup"][i]["TransformType"][:][0]}\n'
-        raise ValueError(msg)
-
-    fixed_params = transform2['TransformFixedParameters'][:]
-    shape = tuple(fixed_params[:3].astype(int))
-    # ITK stores warps in Fortran-order, where the vector components change fastest
-    # Nitransforms expects 3 volumes, not a volume of three-vectors, so transpose
-    warp = np.reshape(
-        transform2['TransformParameters'],
-        (3, *shape),
-        order='F',
-    ).transpose(1, 2, 3, 0)
-
-    warp_affine = np.eye(4)
-    warp_affine[:3, :3] = fixed_params[9:].reshape((3, 3))
-    warp_affine[:3, 3] = fixed_params[3:6]
-    lps_to_ras = np.eye(4) * np.array([-1, -1, 1, 1])
-    warp_affine = lps_to_ras @ warp_affine
-    transforms.insert(0, nt.DenseFieldTransform(nb.Nifti1Image(warp, warp_affine)))
-    return nt.TransformChain(transforms)
 
 
 class ResampleSeriesInputSpec(TraitedSpec):
@@ -788,7 +698,7 @@ def reconstruct_fieldmap(
     )
 
     if not direct:
-        fmap_img = transforms.apply(fmap_img, reference=target)
+        fmap_img = nt.apply(fmap_img, reference=target)
 
     fmap_img.header.set_intent('estimate', name='fieldmap Hz')
     fmap_img.header.set_data_dtype('float32')
