+++ pyproject.toml
@@ -24,7 +24,7 @@ dependencies = [
     "nipype >= 1.8.5",
     "nireports >= 23.2.0",
     "nitime",
-    "nitransforms >= 23.0.1",
+    "nitransforms >= 24.1.1",
     "niworkflows >= 1.12.1",
     "numpy >= 1.21.0",
     "packaging",
